% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonmeta.R
\name{commonmeta_doi}
\alias{commonmeta_doi}
\title{Generate random DOI string}
\usage{
commonmeta_doi(prefix = "10.59350")
}
\arguments{
\item{prefix}{DOI prefix}
}
\value{
A random DOI string
}
\description{
Generate random DOI string
}
\examples{
commonmeta_doi()
}
