% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonmeta-install.R
\name{commonmeta_install}
\alias{commonmeta_install}
\title{Install latest version of commonmeta}
\usage{
commonmeta_install(os = NULL, arch = "x86_64", force = FALSE)
}
\arguments{
\item{os}{Operating system, one of "Linux", "Windows", "Darwin". Defaults
to current operating system. Will be guessed automatically.}

\item{arch}{Architecture}

\item{force}{Whether to force a re-install/trigger an update.}
}
\description{
Downloads binary from
\href{https://github.com/front-matter/commonmeta}{commonmeta} releases,
and installs in system wide cache.
}
\examples{
\dontrun{
commonmeta_install()
}
}
